Change CountingSampler to accept float instead of double (#2592)
Motivation:

`CountingSampler` is currently using `double` for probability, while [Brave](https://github.com/openzipkin/brave/blob/master/brave/src/main/java/brave/sampler/CountingSampler.java#L60) (where this class was originally forked from) is using `float`. This causes issues when passing in floats like `0.01f`, which is converted to `double` as `0.009999999776482582`. This causes `percent` to be `0`.

Modification:

- Change `CountingSampler` to accept a `float`.

Result:

When passing in the probability of `0.01f`, the percentage is correctly calculated as `1`.

+++ core/src/main/java/com/linecorp/armeria/common/util/CountingSampler.java
@@ -58,8 +58,8 @@ final class CountingSampler<T> implements Sampler<T> {
      * @param probability {@code 0.0} means never sample, {@code 1.0} means always sample.
      *                    Otherwise minimum probability is between {@code 0.01} and {@code 1.0}.
      */
-    static <T> Sampler<T> create(double probability) {
-        final int percent = (int) (probability * 100.0);
+    static <T> Sampler<T> create(float probability) {
+        final int percent = (int) (probability * 100.0f);
         checkArgument(percent >= 0 && percent <= 100,
                       "probability: %s (expected: 0.0 <= probability <= 1.0)", probability);
         if (percent == 0) {
